moving model struct to pkg
