+++ cmd/ocm/gcp/delete-wif-config.go
@@ -7,8 +7,8 @@ import (
 	"log"
 
 	"github.com/openshift-online/ocm-cli/cmd/ocm/gcp/mock"
-	"github.com/openshift-online/ocm-cli/cmd/ocm/gcp/models"
 	"github.com/openshift-online/ocm-cli/pkg/gcp"
+	"github.com/openshift-online/ocm-cli/pkg/models"
 	"github.com/pkg/errors"
 
 	"github.com/spf13/cobra"
